requirements.txt
boto3~=1.17.95
botocore~=1.20.95
python-dotenv~=0.17.1
paho-mqtt~=1.5.1
redis~=3.5.3
pygame~=2.0.1